COMMERCE-5049 commerce-machine-learning-impl Apply new naming pattern to commerce machine learning
